Create labeler.yml
Adding workflow for newly pushed PRs. If PRs have 'WIP' in the title, then that PR will be labeled 'Needs Input' and placed in the 'In Development' column of the project board. All other PRs will be labeled with the 'Ready for Review #1' label and placed under the 'Ready for Review' column of the project board.