Revises explanation of chained relational operators
Python handling of expressions with multiple relational operations is
different from many other languages. The previous example would be
correct for Ruby, JS, C/C++, Java... bunches of others.

This change clearly marks the old content as incorrect (in an effort
to be fully transparent with the students) and directs
students to the new content added later, in an ordering determined
by needing to have an understanding of operator precendence
before presenting the easiest way to understand (roughly) what
Python is doing.

An additional header is also added to separate the description of how
parentheses affect expression evaluation, from a series of example
expressions, none of which use parentheses.

An extra space and a misspelling is also cleaned up in those examples.